Changes to transmit/receive funcitons in DataDistributor. Changed arrays in globals.h to int instead of float for serial communication functionality

+++ robot/globals.h
@@ -13,8 +13,8 @@ extern DigitalIn pb;
 /**
  * Data declarations. Initializations will be in main.cpp
  */
-extern float* obstacles;
+extern int* obstacles;
 extern int trajectoryLength;
-extern float* trajectory;
+extern int* trajectory;
 
 #endif /* GLOBALS_H */
